requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
exceptiongroup==1.1.1
flake8==6.0.0
h11==0.14.0
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
PySocks==1.7.1
python-decouple==3.8
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
whitenoise==6.4.0
wsproto==1.2.0
